backend/pyproject.toml
[project]
name = "all-things-ones"
version = "0.1.0"
description = "A project for experimental image processing"
authors = [
    {name = "Nick Sullivan", email = "nick.dave.sullivan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "opencv-python==4.12.0.88",
    "pillow==11.3.0",
    "python-multipart>=0.0.12",
    "scipy==1.16.1",
    "snakeviz==2.2.2",
    "svgpathtools==1.7.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/all_things_ones"]

[tool.uv]
dev-dependencies = []
